Half AStarSearch done

+++ src/euclidean_function.cpp
@@ -5,9 +5,8 @@ namespace ia {
 EuclideanFunction::EuclideanFunction() {}
 
 double EuclideanFunction::operator()(Position position, Position goal) const {
-  return std::sqrt(
-      std::pow(static_cast<double>(position.GetX() - goal.GetX()), 2) +
-      std::pow(static_cast<double>(position.GetY() - goal.GetY()), 2));
+  return std::sqrt(std::pow(position.GetX() - goal.GetX(), 2) +
+                   std::pow(position.GetY() - goal.GetY(), 2));
 }
 
 }  // namespace ia
