+++ src/rectilinear_function.cpp
@@ -5,8 +5,8 @@ namespace ia {
 RectilinearFunction::RectilinearFunction() {}
 
 double RectilinearFunction::operator()(Position position, Position goal) const {
-  return std::abs(static_cast<double>(position.GetX() - goal.GetX())) +
-         std::abs(static_cast<double>(position.GetY() - goal.GetY()));
+  return std::abs(position.GetX() - goal.GetX()) +
+         std::abs(position.GetY() - goal.GetY());
 }
 
 }  // namespace ia
